+++ resources/views/app.blade.php
@@ -31,11 +31,20 @@
             }
         </style>
 
-        <title inertia>{{ config('app.name', 'Laravel') }}</title>
+        <title inertia>{{ config('app.name', 'Store') }}</title>
 
-        <link rel="icon" href="/favicon.ico" sizes="any">
-        <link rel="icon" href="/favicon.svg" type="image/svg+xml">
-        <link rel="apple-touch-icon" href="/apple-touch-icon.png">
+        @php
+            $storeFavicon = app(\App\Services\StoreSettingService::class)->getSetting('store_favicon');
+        @endphp
+
+        @if($storeFavicon)
+            <link rel="icon" href="{{ asset('storage/' . $storeFavicon) }}" type="image/png">
+            <link rel="apple-touch-icon" href="{{ asset('storage/' . $storeFavicon) }}">
+        @else
+            <link rel="icon" href="/favicon.ico" sizes="any">
+            <link rel="icon" href="/favicon.svg" type="image/svg+xml">
+            <link rel="apple-touch-icon" href="/apple-touch-icon.png">
+        @endif
 
         <link rel="preconnect" href="https://fonts.bunny.net">
         <link href="https://fonts.bunny.net/css?family=instrument-sans:400,500,600" rel="stylesheet" />